+++ src/main/java/com/gyeryongbrother/pickandtest/domain/service/ports/input/PortfolioQueryService.java
@@ -1,9 +1,12 @@
 package com.gyeryongbrother.pickandtest.domain.service.ports.input;
 
+import com.gyeryongbrother.pickandtest.domain.service.dto.PortfolioResponse;
 import com.gyeryongbrother.pickandtest.domain.service.dto.PortfolioStockResponse;
 import java.util.List;
 
-public interface PortfolioStockQueryService {
+public interface PortfolioQueryService {
 
     List<PortfolioStockResponse> findAllByPortfolioId(Long portfolioId);
+
+    List<PortfolioResponse> findAllPortfolios();
 }
