+++ src/main/java/com/gyeryongbrother/pickandtest/member/domain/service/ports/input/MemberService.java
@@ -1,13 +1,9 @@
 package com.gyeryongbrother.pickandtest.member.domain.service.ports.input;
 
-import com.gyeryongbrother.pickandtest.domain.service.dto.PortfolioResponse;
 import com.gyeryongbrother.pickandtest.member.domain.service.dto.RegisterMemberCommand;
 import com.gyeryongbrother.pickandtest.member.domain.service.dto.RegisterMemberResponse;
-import java.util.List;
 
 public interface MemberService {
 
     RegisterMemberResponse register(RegisterMemberCommand registerMemberCommand);
-
-    List<PortfolioResponse> findAllPortfolios(Long memberId);
 }
